Add a root package and move classes
