+++ src/main/java/de/fadinaoum/obs_to_op/Subscription.java
@@ -1,4 +1,4 @@
-package de.fadinaoum;
+package de.fadinaoum.obs_to_op;
 
 public interface Subscription {
     void cancel();
