+++ src/main/java/de/fadinaoum/obs_to_op/observer/Observer.java
@@ -1,4 +1,4 @@
-package de.fadinaoum.observer;
+package de.fadinaoum.obs_to_op.observer;
 
 public interface Observer<T> {
     void onNext(T value);
